+++ lib/beacon/live_admin/web.ex
@@ -24,7 +24,7 @@ defmodule Beacon.LiveAdmin.Web do
         layouts: [html: Beacon.LiveAdmin.Layouts]
 
       import Plug.Conn
-      import Beacon.LiveAdmin.Gettext
+      use Gettext, backend: Beacon.LiveAdmin.Gettext
     end
   end
 
@@ -61,7 +61,7 @@ defmodule Beacon.LiveAdmin.Web do
       import Phoenix.HTML
       import Beacon.LiveAdmin.AdminComponents
       import Beacon.LiveAdmin.Components, only: [template_error: 1]
-      import Beacon.LiveAdmin.Gettext
+      use Gettext, backend: Beacon.LiveAdmin.Gettext
       import LiveSvelte
 
       import Beacon.LiveAdmin.Router,